backend/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
Automat==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
distlib==0.3.4
feedparser==6.0.8
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JSON==0.3.4
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Twisted==0.1.2
flast==1.3
greenlet==1.1.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
observable==1.0.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytrends==4.8.0
pytz==2021.3
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
tomli==2.0.1
Twisted==22.2.0
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.13.2
Werkzeug==2.0.3
zope.interface==5.4.0
